ZFIN 7891: Fix blank fish names (#188)
* ZFIN-7891

Updating queries to include checks for empty names.
Key the fixes in GenotypeNamingIssues by both the original name and the computed name.
Add Yvonne and Ryan T. to recipient list for Check-Genotype-Name report and Check-Fish-Name report failure notifications.
Add confirmation before deleting feature marker relationship.
Broaden the scope of the report that checks for incorrect fish names.
Add liquibase for fixing fish and genotype names for known issues.
get_fish_name uses current genotype name. Updated to use get_genotype_display() function if current genotype name is blank.

+++ source/org/zfin/gwt/curation/ui/feature/FeatureRelationshipPresenter.java
@@ -2,6 +2,7 @@ package org.zfin.gwt.curation.ui.feature;
 
 import com.google.gwt.event.dom.client.ClickEvent;
 import com.google.gwt.event.dom.client.ClickHandler;
+import com.google.gwt.user.client.Window;
 import com.google.gwt.user.client.ui.Button;
 import org.zfin.gwt.curation.event.CurationEvent;
 import org.zfin.gwt.curation.event.EventType;
@@ -367,6 +368,9 @@ public class FeatureRelationshipPresenter implements HandlesError {
             addClickHandler(new ClickHandler() {
                 @Override
                 public void onClick(ClickEvent event) {
+                    if (!Window.confirm("Are you sure you want to delete this relationship?")) {
+                        return;
+                    }
                     setEnabled(false);
                     FeatureRPCService.App.getInstance().deleteFeatureMarkerRelationship(featureMarkerRelationshipDTO,
                             new FeatureEditCallBack<Void>("Unable to remove feature marker relationship: " + featureMarkerRelationshipDTO, handlesError) {
